+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day07/p019/Solution1005_02.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day07.p019;
+package com.noob.algorithm.daily.archive.plan02.day07.p019;
 
 import java.util.Arrays;
 import java.util.Comparator;
