+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day07/p019/Solution860_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day07.p019;
+package com.noob.algorithm.daily.archive.plan02.day07.p019;
 
 /**
  * 🟢 860 柠檬水找零 - https://leetcode.cn/problems/lemonade-change/description/
