+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day07/p021/Solution121_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day07.p021;
+package com.noob.algorithm.daily.archive.plan02.day07.p021;
 
 /**
  * 🟡 121 买卖股票的最佳时机I - https://leetcode.cn/problems/best-time-to-buy-and-sell-stock/description/
