+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day07/p022/Solution435_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day07.p022;
+package com.noob.algorithm.daily.archive.plan02.day07.p022;
 
 import java.util.Arrays;
 import java.util.Comparator;
