+++ rebirth/base/KeepJava/src/com/noob/algorithm/daily/archive/plan02/day07/p024/Solution968_01.java
@@ -1,4 +1,4 @@
-package com.noob.algorithm.daily.plan02.day07.p024;
+package com.noob.algorithm.daily.archive.plan02.day07.p024;
 
 /**
  * 🔴 968 监控二叉树 - https://leetcode.cn/problems/binary-tree-cameras/description/
